core: make the app work
