+++ gsl/urls.py
@@ -19,7 +19,7 @@ urlpatterns = [
     path("oidc/", include("mozilla_django_oidc.urls")),
     path("projets/", include(("gsl_projet.urls", "gsl_projet"), "projet")),
     path(
-        "programmation/",
-        include(("gsl_programmation.urls", "gsl_programmation"), "programmation"),
+        "simulation/",
+        include(("gsl_simulation.urls", "gsl_simulation"), "simulation"),
     ),
 ]
