+++ gsl_core/templates/blocks/main_menu.html
@@ -31,7 +31,7 @@
                        {% if request.path == projet_url %}aria-current="true"{% endif %}>Projets 2025</a>
                 </li>
                 <li class="fr-nav__item">
-                    {% url 'programmation:simulation-list' as simulations_url %}
+                    {% url 'simulation:simulation_list' as simulations_url %}
                     <a class="fr-nav__link"
                        href="{{ simulations_url }}"
                        target="_self"
