+++ gsl_programmation/apps.py
@@ -4,4 +4,4 @@ from django.apps import AppConfig
 class GslProgrammationConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "gsl_programmation"
-    verbose_name = "4. Programmation"
+    verbose_name = "5. Programmation"
