--- gsl_programmation/templates/includes/_simulation_projet_list_filters.html
@@ -1,115 +1,0 @@
-<form method="get"
-      action="{{ request.path }}{% querystring %}"
-      class="fr-background-alt--blue-france">
-
-    <div class="projets-filters-layout fr-container--fluid">
-        <div class="fr-grid-row fr-grid-row--gutters">
-            <div class="fr-col-12 fr-col-md-6 fr-col-lg-3">
-                <label class="fr-label" for="tri">
-                    Trier par
-                </label>
-                <select name="tri" id="tri" class="fr-select" onchange="this.form.submit()">
-                    <option value=""
-                            {% if request.GET.tri == None %}selected{% endif %}
-                            disabled
-                            hidden>
-                        Tri
-                    </option>
-                    <optgroup label="Date">
-                        <option value="date_desc"
-                                {% if request.GET.tri == 'date_desc' %}selected{% endif %}>
-                            Les plus récentes en premier
-                        </option>
-                        <option value="date_asc"
-                                {% if request.GET.tri == 'date_asc' %}selected{% endif %}>
-                            Les plus anciennes en premier
-                        </option>
-                    </optgroup>
-                    <optgroup label="Coût total">
-                        <option value="cout_asc"
-                                {% if request.GET.tri == 'cout_asc' %}selected{% endif %}>
-                            Croissant
-                        </option>
-                        <option value="cout_desc"
-                                {% if request.GET.tri == 'cout_desc' %}selected{% endif %}>
-                            Décroissant
-                        </option>
-                    </optgroup>
-                    <optgroup label="Commune">
-                        <option value="commune_asc"
-                                {% if request.GET.tri == 'commune_asc' %}selected{% endif %}>
-                            A → Z
-                        </option>
-                        <option value="commune_desc"
-                                {% if request.GET.tri == 'commune_desc' %}selected{% endif %}>
-                            Z → A
-                        </option>
-                    </optgroup>
-                </select>
-            </div>
-            <div class="fr-col-12 fr-col-md-6 fr-col-lg-3">
-                <label class="fr-label" for="dispositif">
-                    Dispositif
-                </label>
-                <select name="dispositif"
-                        id="dispositif"
-                        class="fr-select"
-                        onchange="this.form.submit()">
-                    <option value="">
-                        Tous les dispositifs
-                    </option>
-                    <option value="DETR"
-                            {% if request.GET.dispositif == 'DETR' %}selected{% endif %}>
-                        DETR
-                    </option>
-                    <option value="DSIL"
-                            {% if request.GET.dispositif == 'DSIL' %}selected{% endif %}>
-                        DSIL
-                    </option>
-                </select>
-            </div>
-            <div class="fr-col-12 fr-col-md-6 fr-col-lg-3">
-                <label class="fr-label" for="porteur">
-                    Porteurs de projet
-                </label>
-                <select name="porteur"
-                        id="porteur"
-                        class="fr-select"
-                        onchange="this.form.submit()">
-                    <option value="">
-                        Tous les porteurs
-                    </option>
-                    {% for value in porteur_mappings %}
-                        <option value={{ value }} {% if request.GET.porteur == value %}selected{% endif %}>
-                            {{ value }}
-                        </option>
-                    {% endfor %}
-                </select>
-            </div>
-            <div class="fr-col-12 fr-col-md-6 fr-col-lg-3">
-                <label class="fr-label" for="cout_min">
-                    Coût total minimum
-                </label>
-                <input type="number"
-                       name="cout_min"
-                       id="cout_min"
-                       class="fr-input"
-                       placeholder="Montant minimum"
-                       onfocusout="this.form.submit()"
-                       value="{{ request.GET.cout_min }}">
-            </div>
-            <div class="fr-col-12 fr-col-md-6 fr-col-lg-3">
-                <label class="fr-label" for="cout_max">
-                    Coût total maximum
-                </label>
-                <input type="number"
-                       name="cout_max"
-                       id="cout_max"
-                       class="fr-input"
-                       placeholder="Montant maximum"
-                       onfocusout="this.form.submit()"
-                       value="{{ request.GET.cout_max }}">
-            </div>
-        </div>
-    </div>
-</form>
