+++ gsl_programmation/tests/services/test_simulation_services.py
@@ -8,12 +8,12 @@ from gsl_core.tests.factories import (
     PerimetreRegionalFactory,
 )
 from gsl_programmation.models import Enveloppe, Simulation
-from gsl_programmation.services.simulation_service import SimulationService
 from gsl_programmation.tests.factories import (
     DetrEnveloppeFactory,
     DsilEnveloppeFactory,
     SimulationFactory,
 )
+from gsl_simulation.services.simulation_service import SimulationService
 
 pytestmark = pytest.mark.django_db
