+++ gsl_simulation/apps.py
@@ -4,3 +4,4 @@ from django.apps import AppConfig
 class GslSimulationConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "gsl_simulation"
+    verbose_name = "4. Simulation"
