+++ gsl_simulation/models.py
@@ -27,7 +27,7 @@ class Simulation(models.Model):
     def get_absolute_url(self):
         from django.urls import reverse
 
-        return reverse("programmation:simulation_detail", kwargs={"slug": self.slug})
+        return reverse("simulation:simulation_detail", kwargs={"slug": self.slug})
 
     def get_projet_status_summary(self):
         default_status_summary = {
