+++ gsl_simulation/templates/htmx/projet_update.html
