+++ gsl_simulation/templates/includes/_enveloppe_summary.html
