+++ gsl_simulation/templates/includes/_montant_form.html
@@ -1,17 +1,17 @@
-<form action="{% url 'programmation:patch-simulation-projet-taux' simu.id %}"
+<form action="{% url 'simulation:patch-simulation-projet-montant' simu.id %}"
       method="POST"
       hx-on="submit: event.preventDefault()">
     {% csrf_token %}
-    <div class="gsl-projet-table__input-wrap gsl-projet-table__rate">
-        <input name="taux"
+    <div div class="gsl-projet-table__input-wrap">
+        <input name="montant"
                type="text"
                inputmode="numeric"
                class="fr-input input-simulation-projet-{{ simu.pk }}"
-               hx-patch="{% url 'programmation:patch-simulation-projet-taux' simu.id %}"
+               hx-patch="{% url 'simulation:patch-simulation-projet-montant' simu.id %}"
                hx-disabled-elt=".input-simulation-projet-{{ simu.pk }}"
                hx-trigger="change, keyup[key=='Enter']"
-               value="{{ simu.taux }}">
-        <span class="gsl-projet-table__input-symbol">%</span>
+               value="{{ simu.montant|floatformat:"2g" }}">
+        <span class="gsl-projet-table__input-symbol">€</span>
     </div>
     <input type="hidden" name="filter_params" value="{{ filter_params }}">
 </form>
