+++ gsl_simulation/templates/includes/_simulation_detail_row.html
