+++ gsl_simulation/templates/includes/_status_form.html
@@ -1,11 +1,11 @@
 {% load gsl_filters %}
-<form action="{% url 'programmation:patch-simulation-projet-status' simu.id %}"
+<form action="{% url 'simulation:patch-simulation-projet-status' simu.id %}"
       method="POST">
     {% csrf_token %}
     <select class="fr-select gsl-projet-table__select input-simulation-projet-{{ simu.pk }}"
             id="status-select"
             name="status"
-            hx-patch="{% url 'programmation:patch-simulation-projet-status' simu.id %}"
+            hx-patch="{% url 'simulation:patch-simulation-projet-status' simu.id %}"
             hx-disabled-elt="this">
         {% for key, value in available_states %}
             <option value={{ key }} {% if simu.status == key %} selected {% endif %}>
