+++ gsl_simulation/templates/includes/_status_summary.html
